Consistency in the switch options
Ensuring the switch options are all size based so the iPhone/iPad prefix could even be removed (though I’m not advocating that).

I’ve updated comments, README, and other documentation as well as the example app to reflect the library changes.

This addresses Issue #1

Project files all updated for Xcode 9, and pods 1.3.1.